cupe-kg-backend/requirements.txt
flask
flask-cors
pymongo
neo4j
transformers
torch
scikit-learn
numpy
scipy
python-dotenv
pandas